Fixed 64 bit types usage in Javascript, was broken on 32 bit machines.
git-svn-id: http://yate.null.ro/svn/yate/trunk@5612 acf43c95-373e-0410-b603-e72c3f656dc1

+++ libs/yscript/jsobjects.cpp
@@ -537,7 +537,7 @@ bool JsArray::runField(ObjList& stack, const ExpOperation& oper, GenObject* cont
 	oper.name().c_str(),toString().c_str(),this);
     if (oper.name() == YSTRING("length")) {
 	// Reflects the number of elements in an array.
-	ExpEvaluator::pushOne(stack,new ExpOperation(length()));
+	ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)length()));
 	return true;
     }
     return JsObject::runField(stack,oper,context);
@@ -556,7 +556,7 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 	    const_cast<String&>(op->name()) = (unsigned int)m_length++;
 	    params().addParam(op);
 	}
-	ExpEvaluator::pushOne(stack,new ExpOperation(length()));
+	ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)length()));
     }
     else if (oper.name() == YSTRING("pop")) {
 	// Removes the last element from an array and returns that element
@@ -597,18 +597,18 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 	    return false;
 	JsArray* array = new JsArray(mutex());
 	// copy this array
-	for (long int i = 0; i < m_length; i++)
+	for (int32_t i = 0; i < m_length; i++)
 	    array->params().addParam(params().getParam(String(i)));
 	array->setLength(length());
 	// add parameters (JsArray of JsObject)
-	for (long int i = oper.number(); i; i--) {
+	for (int32_t i = oper.number(); i; i--) {
 	    ExpOperation* op = popValue(stack,context);
 	    ExpWrapper* obj = YOBJECT(ExpWrapper,op);
 	    if (!obj)
 		continue;
 	    JsArray* ja = (JsArray*)obj->getObject(YATOM("JsArray"));
 	    if (ja) {
-		for (long int i = 0; i < ja->length(); i++)
+		for (int32_t i = 0; i < ja->length(); i++)
 		    array->params().addParam(String(i + array->length()),ja->params().getValue(String(i)));
 		array->setLength(array->length() + ja->length());
 	    }
@@ -639,7 +639,7 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 	    TelEngine::destruct(op);
 	}
 	String result;
-	for (long int i = 0; i < length(); i++)
+	for (int32_t i = 0; i < length(); i++)
 	    result.append(params()[String(i)],separator);
 	ExpEvaluator::pushOne(stack,new ExpOperation(result));
     }
@@ -650,10 +650,10 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 	NamedList reversed("");
 	String separator = ",";
 	String toCopy;
-	for (long int i = 0; i < length(); i++)
+	for (int32_t i = 0; i < length(); i++)
 	    toCopy.append(params()[String(i)],separator);
 	reversed.copyParams(params(),toCopy);
-	for (long int i = length(); i; i--)
+	for (int32_t i = length(); i; i--)
 	    params().setParam(String(length() - i),reversed.getValue(String(i - 1)));
     }
     else if (oper.name() == YSTRING("shift")) {
@@ -673,7 +673,7 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 	else {
 	    ExpEvaluator::pushOne(stack,new ExpOperation(params().getValue("0")));
 	    // shift : value n+1 becomes value n
-	    for (long int i = 0; i < length() - 1; i++)
+	    for (int32_t i = 0; i < length() - 1; i++)
 		params().setParam(String(i),params().getValue(String(i + 1)));
 	    params().clearParam(String(length() - 1));
 	    setLength(length() - 1);
@@ -691,10 +691,10 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 	// myFish after: ["drum", "lion", "angel", "clown"]
 	// New length: 4
 	// shift array
-	long shift = oper.number();
-	for (long int i = length(); i; i--)
+	int32_t shift = oper.number();
+	for (int32_t i = length(); i; i--)
 	    params().setParam(String(i - 1 + shift),params().getValue(String(i - 1)));
-	for (long int i = shift; i; i--) {
+	for (int32_t i = shift; i; i--) {
 	    ExpOperation* op = popValue(stack,context);
 	    ExpWrapper* obj = YOBJECT(ExpWrapper,op);
 	    if (!obj)
@@ -708,7 +708,7 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 	    TelEngine::destruct(op);
 	}
 	setLength(length() + shift);
-	ExpEvaluator::pushOne(stack,new ExpOperation(length()));
+	ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)length()));
     }
     else if (oper.name() == YSTRING("slice"))
 	return runNativeSlice(stack,oper,context);
@@ -723,7 +723,7 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 	// var myVar = monthNames.toString(); // assigns "Jan,Feb,Mar,Apr" to myVar.
 	String separator = ",";
 	String result;
-	for (long int i = 0; i < length(); i++)
+	for (int32_t i = 0; i < length(); i++)
 	    result.append(params()[String(i)],separator);
 	ExpEvaluator::pushOne(stack,new ExpOperation(result));
     } else if (oper.name() == YSTRING("indexOf")) {
@@ -964,7 +964,7 @@ bool JsMath::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
     if (oper.name() == YSTRING("abs")) {
 	if (!oper.number())
 	    return false;
-	long int n = 0;
+	int64_t n = 0;
 	for (long int i = oper.number(); i; i--) {
 	    ExpOperation* op = popValue(stack,context);
 	    if (op->isInteger())
@@ -978,7 +978,7 @@ bool JsMath::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
     else if (oper.name() == YSTRING("max")) {
 	if (!oper.number())
 	    return false;
-	long int n = LONG_MIN;
+	int64_t n = LLONG_MIN;
 	for (long int i = oper.number(); i; i--) {
 	    ExpOperation* op = popValue(stack,context);
 	    if (op->isInteger() && op->number() > n)
@@ -990,7 +990,7 @@ bool JsMath::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
     else if (oper.name() == YSTRING("min")) {
 	if (!oper.number())
 	    return false;
-	long int n = LONG_MAX;
+	int64_t n = LLONG_MAX;
 	for (long int i = oper.number(); i; i--) {
 	    ExpOperation* op = popValue(stack,context);
 	    if (op->isInteger() && op->number() < n)
@@ -1021,7 +1021,7 @@ bool JsMath::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
 	if (min != 0)
 	    interval -= min + 1;
 	long rand = (Random::random() % interval) + min;
-	ExpEvaluator::pushOne(stack,new ExpOperation((long)rand));
+	ExpEvaluator::pushOne(stack,new ExpOperation(rand));
     }
     else
 	return JsObject::runNative(stack,oper,context);
@@ -1081,7 +1081,7 @@ bool JsDate::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
 	oper.name().c_str(),toString().c_str(),this);
     if (oper.name() == YSTRING("now")) {
 	// Returns the number of milliseconds elapsed since 1 January 1970 00:00:00 UTC.
-	ExpEvaluator::pushOne(stack,new ExpOperation((long int)Time::msecNow()));  // should check conversion from u_int64_t 
+	ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)Time::msecNow()));
     }
     else if (oper.name() == YSTRING("getDate")) {
 	// Returns the day of the month for the specified date according to local time.
@@ -1089,7 +1089,7 @@ bool JsDate::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
 	int year = 0;
 	unsigned int month = 0, day = 0, hour = 0, minute = 0, sec = 0;
 	if (Time::toDateTime(m_time + m_offs,year,month,day,hour,minute,sec))
-	    ExpEvaluator::pushOne(stack,new ExpOperation((long int)day));
+	    ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)day));
 	else
 	    return false;
     }
@@ -1098,7 +1098,7 @@ bool JsDate::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
 	int year = 0;
 	unsigned int month = 0, day = 0, hour = 0, minute = 0, sec = 0, wday = 0;
 	if (Time::toDateTime(m_time + m_offs,year,month,day,hour,minute,sec,&wday))
-	    ExpEvaluator::pushOne(stack,new ExpOperation((long int)wday));
+	    ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)wday));
 	else
 	    return false;
     }
@@ -1107,7 +1107,7 @@ bool JsDate::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
 	int year = 0;
 	unsigned int month = 0, day = 0, hour = 0, minute = 0, sec = 0;
 	if (Time::toDateTime(m_time + m_offs,year,month,day,hour,minute,sec))
-	    ExpEvaluator::pushOne(stack,new ExpOperation((long int)year));
+	    ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)year));
 	else
 	    return false;
     }
@@ -1116,20 +1116,20 @@ bool JsDate::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
 	int year = 0;
 	unsigned int month = 0, day = 0, hour = 0, minute = 0, sec = 0;
 	if (Time::toDateTime(m_time + m_offs,year,month,day,hour,minute,sec))
-	    ExpEvaluator::pushOne(stack,new ExpOperation((long int)hour));
+	    ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)hour));
 	else
 	    return false;
     }
     else if (oper.name() == YSTRING("getMilliseconds")) {
 	// Returns just the milliseconds part ( 0 - 999 )
-	ExpEvaluator::pushOne(stack,new ExpOperation((long int)m_msec));
+	ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)m_msec));
     }
     else if (oper.name() == YSTRING("getMinutes")) {
 	// Returns the minute ( 0 - 59 ) of the specified date according to local time.
 	int year = 0;
 	unsigned int month = 0, day = 0, hour = 0, minute = 0, sec = 0;
 	if (Time::toDateTime(m_time + m_offs,year,month,day,hour,minute,sec))
-	    ExpEvaluator::pushOne(stack,new ExpOperation((long int)minute));
+	    ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)minute));
 	else
 	    return false;
     }
@@ -1138,7 +1138,7 @@ bool JsDate::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
 	int year = 0;
 	unsigned int month = 0, day = 0, hour = 0, minute = 0, sec = 0;
 	if (Time::toDateTime(m_time + m_offs,year,month,day,hour,minute,sec))
-	    ExpEvaluator::pushOne(stack,new ExpOperation((long int)month - 1));
+	    ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)month - 1));
 	else
 	    return false;
     }
@@ -1147,17 +1147,17 @@ bool JsDate::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
 	int year = 0;
 	unsigned int month = 0, day = 0, hour = 0, minute = 0, sec = 0;
 	if (Time::toDateTime(m_time + m_offs,year,month,day,hour,minute,sec))
-	    ExpEvaluator::pushOne(stack,new ExpOperation((long int)sec));
+	    ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)sec));
 	else
 	    return false;
     }
     else if (oper.name() == YSTRING("getTime")) {
 	// Returns the time in milliseconds since UNIX Epoch
-	ExpEvaluator::pushOne(stack,new ExpOperation(1000 * ((long int)m_time) + (long int)m_msec));
+	ExpEvaluator::pushOne(stack,new ExpOperation(1000 * ((int64_t)m_time) + (int64_t)m_msec));
     }
     else if (oper.name() == YSTRING("getTimezoneOffset")) {
 	// Returns the UTC to local difference in minutes, positive goes west
-	ExpEvaluator::pushOne(stack,new ExpOperation((long int)(m_offs / -60)));
+	ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)(m_offs / -60)));
     }
     else if (oper.name() == YSTRING("getUTCDate")) {
 	// Returns the day of the month for the specified date according to local time.
@@ -1165,7 +1165,7 @@ bool JsDate::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
 	int year = 0;
 	unsigned int month = 0, day = 0, hour = 0, minute = 0, sec = 0;
 	if (Time::toDateTime(m_time,year,month,day,hour,minute,sec))
-	    ExpEvaluator::pushOne(stack,new ExpOperation((long int)day));
+	    ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)day));
 	else
 	    return false;
     }
@@ -1174,7 +1174,7 @@ bool JsDate::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
 	int year = 0;
 	unsigned int month = 0, day = 0, hour = 0, minute = 0, sec = 0, wday = 0;
 	if (Time::toDateTime(m_time,year,month,day,hour,minute,sec,&wday))
-	    ExpEvaluator::pushOne(stack,new ExpOperation((long int)wday));
+	    ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)wday));
 	else
 	    return false;
     }
@@ -1183,7 +1183,7 @@ bool JsDate::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
 	int year = 0;
 	unsigned int month = 0, day = 0, hour = 0, minute = 0, sec = 0;
 	if (Time::toDateTime(m_time,year,month,day,hour,minute,sec))
-	    ExpEvaluator::pushOne(stack,new ExpOperation((long int)year));
+	    ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)year));
 	else
 	    return false;
     }
@@ -1192,20 +1192,20 @@ bool JsDate::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
 	int year = 0;
 	unsigned int month = 0, day = 0, hour = 0, minute = 0, sec = 0;
 	if (Time::toDateTime(m_time,year,month,day,hour,minute,sec))
-	    ExpEvaluator::pushOne(stack,new ExpOperation((long int)hour));
+	    ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)hour));
 	else
 	    return false;
     }
     else if (oper.name() == YSTRING("getUTCMilliseconds")) {
 	// Returns just the milliseconds part ( 0 - 999 )
-	ExpEvaluator::pushOne(stack,new ExpOperation((long int)m_msec));
+	ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)m_msec));
     }
     else if (oper.name() == YSTRING("getUTCMinutes")) {
 	// Returns the minute ( 0 - 59 ) of the specified date according to local time.
 	int year = 0;
 	unsigned int month = 0, day = 0, hour = 0, minute = 0, sec = 0;
 	if (Time::toDateTime(m_time,year,month,day,hour,minute,sec))
-	    ExpEvaluator::pushOne(stack,new ExpOperation((long int)minute));
+	    ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)minute));
 	else
 	    return false;
     }
@@ -1214,7 +1214,7 @@ bool JsDate::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
 	int year = 0;
 	unsigned int month = 0, day = 0, hour = 0, minute = 0, sec = 0;
 	if (Time::toDateTime(m_time,year,month,day,hour,minute,sec))
-	    ExpEvaluator::pushOne(stack,new ExpOperation((long int)month - 1));
+	    ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)month - 1));
 	else
 	    return false;
     }
@@ -1223,7 +1223,7 @@ bool JsDate::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
 	int year = 0;
 	unsigned int month = 0, day = 0, hour = 0, minute = 0, sec = 0;
 	if (Time::toDateTime(m_time,year,month,day,hour,minute,sec))
-	    ExpEvaluator::pushOne(stack,new ExpOperation((long int)sec));
+	    ExpEvaluator::pushOne(stack,new ExpOperation((int64_t)sec));
 	else
 	    return false;
     }
